mvp/pyproject.toml
[project]
name = "pauseai-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "discord-py>=2.6.4",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
